--- src/pirri/templates/login.html
@@ -1,64 +1,0 @@
-<!doctype html>
-<html ng-app="pirrilogin">
-<head>
-        <title>PirriGo - Login</title>
-        <link rel="icon" href="/static/icon.png">
-        <link rel="stylesheet" href="/static/bower_components/bootstrap/dist/css/bootstrap.css">
-        <link rel="stylesheet" href="/static/bower_components/ng-responsive-calendar/dist/css/calendar.min.css" />
-        <link rel="stylesheet" href="/static/style.css">
-        <script src="/static/bower_components/jquery/dist/jquery.min.js"></script>
-        <script src="/static/bower_components/angular/angular.js"></script>
-        <script type="text/javascript" src="/static/bower_components/chart.js/dist/Chart.js"></script>
-        <script type="text/javascript" src="/static/bower_components/moment/moment.js"></script>
-        <script src="/static/bower_components/ng-responsive-calendar/dist/js/calendar-tpls.min.js"></script>
-        <script type="text/javascript" src="/static/bower_components/angular-chart.js/dist/angular-chart.js"></script>
-        <script src="/static/bower_components/angular-moment/angular-moment.js"></script>
-        <script src="/static/bower_components/bootstrap/dist/js/bootstrap.js"></script>
-        <script type="text/javascript" src="/static/bower_components/angular-cookies/angular-cookies.js"></script>
-        <script src="/static/bower_components/angular-animate/angular-animate.js"></script>
-        <script src="/static/bower_components/angular-sanitize/angular-sanitize.js"></script>
-        <script src="/static/bower_components/angular-route/angular-route.js"></script>
-        <script type="text/javascript" src="/static/bower_components/prefix/src/prefix.js"></script>
-        <script type="text/javascript" src="/static/bower_components/angular-messages/angular-messages.js"></script>
-
-        <script src="/static/bower_components/angular-base64/angular-base64.js"></script>
-        <script src="/static/login.js"></script>
-</head>
-
-<body class="text-shadow">
-        <div ng-controller="PirriLoginControl as plc">
-                <div class="login pfield">
-                        <h1 id="pirrilogo">
-                                <img height="100" width="150" src="static/icon_text.png">
-                        </h1>
-                        <br/>
-                        <div ng-show="metadata.banner != undefined && metadata.banner != ''">
-                                <h4 style="text-align: center">
-                                        {{ metadata.banner }}
-                                </h4>
-                        </div>
-                        <form>
-                                <input type="text" name="u" placeholder="Username" required="required" ng-model="username">
-                                <input type="password" name="p" placeholder="Password" required="required" ng-model="password">
-                                <button ng-show="username != undefined && password != undefined && username != '' && password != ''" ng-click="plc.Login(username, password)"
-                                        class="btn btn-primary btn-block btn-large">Log In</button>
-                        </form>
-                        <div style="padding-top: 40px">
-                                <a rel="license" href="http://creativecommons.org/licenses/by-nc/4.0/">
-                                        <img alt="Creative Commons License" style="border-width:0" src="https://i.creativecommons.org/l/by-nc/4.0/80x15.png"
-                                        />
-                                </a>
-                                <br />
-                                <span xmlns:dct="http://purl.org/dc/terms/" property="dct:title">PirriGo</span> by
-                                <a xmlns:cc="http://creativecommons.org/ns#" href="http://vacovsky.us"
-                                        property="cc:attributionName" rel="cc:attributionURL">Joe Vacovsky Jr.</a> is licensed under a
-                                <a rel="license" href="http://creativecommons.org/licenses/by-nc/4.0/">Creative Commons Attribution-NonCommercial 4.0 International License</a>.
-                                <br />Based on a work at
-                                <a xmlns:dct="http://purl.org/dc/terms/" href="https://github.com/vacoj/pirrigo"
-                                        rel="dct:source">https://github.com/vacoj/pirrigo</a>.
-                        </div>
-                </div>
-        </div>
-</body>
-
-</html>
