+++ src/schema.go
@@ -2,17 +2,19 @@ package main
 
 import (
 	"./data"
+	"./pirri"
+	"./settings"
 )
 
 func migrateDataSchema() {
 	d := data.Service()
-	d.db.AutoMigrate(
-		&Station{},
-		&DripNode{},
-		&GpioPin{},
-		&StationHistory{},
-		&StationSchedule{},
-		&Settings{},
+	d.DB.AutoMigrate(
+		&pirri.Station{},
+		&pirri.DripNode{},
+		&pirri.GpioPin{},
+		&pirri.StationHistory{},
+		&pirri.StationSchedule{},
+		&settings.Settings{},
 	)
 }
 
@@ -21,7 +23,7 @@ func firstRunDBSetup() {
 	d := data.Service()
 	gpios := []int{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28}
 	for pin := range gpios {
-		d.db.Create(&GpioPin{
+		d.DB.Create(&pirri.GpioPin{
 			GPIO:   pin,
 			Notes:  "",
 			Common: false,
