+++ src/static/bower.json
@@ -1,0 +1,48 @@
+{
+  "name": "pirrigo",
+  "homepage": "https://github.com/vacoj/pirrigo",
+  "authors": [
+    "Joe Vacovsky <vacovsky@gmail.com>"
+  ],
+  "description": "Manager for home garden irrigation systems",
+  "main": "main.go",
+  "keywords": [
+    "irrigation"
+  ],
+  "license": "?",
+  "private": true,
+  "ignore": [
+    "**/.*",
+    "node_modules",
+    "bower_components",
+    "test",
+    "tests"
+  ],
+  "dependencies": {
+    "angular-animate": "^1.6.2",
+    "angular-bootstrap": "^2.5.0",
+    "angular-aria": "^1.6.2",
+    "angular-chart.js": "^1.1.1",
+    "angular-chart": "^0.5.0",
+    "angular-cookies": "^1.6.2",
+    "angular-material": "^1.1.3",
+    "angular-messages": "^1.6.2",
+    "angular-moment": "^1.0.1",
+    "angular-route": "^1.6.2",
+    "angular-sanitize": "^1.6.2",
+    "angular": "^1.6.2",
+    "bootstrap": "^3.3.7",
+    "chart.js": "chartjs#^2.5.0",
+    "jquery-ui": "^1.12.1",
+    "jquery": "^3.1.1",
+    "moment": "^2.17.1",
+    "ui-bootstrap": "^2.5.0",
+    "ng-responsive-calendar": "^0.1.16",
+    "prefix": "^1.0.3",
+    "angular-base64": "^2.0.5"
+  },
+  "resolutions": {
+    "angular": "^1.6.2",
+    "bootstrap": "^3.3.7"
+  }
+}
